+++ Common/AuthenticationType.cs
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) 2020, Oracle and/or its affiliates.
+ * This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
+ */
+
+namespace Oci.PSModules.Common.Cmdlets
+{
+    public enum AuthenticationType
+    {
+        ApiKey,
+        InstancePrincipal
+    }
+}