+++ Database/Cmdlets/Move-OCIDatabaseAutonomousExadataInfrastructureCompartment.cs
@@ -20,8 +20,8 @@ namespace Oci.DatabaseService.Cmdlets
     public class MoveOCIDatabaseAutonomousExadataInfrastructureCompartment : OCIDatabaseCmdlet
     {
         
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = @"Request to move Autonomous Exadata Infrastructure to a different compartment", ParameterSetName = StatusParamSet)]
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = @"Request to move Autonomous Exadata Infrastructure to a different compartment", ParameterSetName = Default)]
+        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = @"Request to move an Autonomous Exadata Infrastructure resource to a different compartment.", ParameterSetName = StatusParamSet)]
+        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = @"Request to move an Autonomous Exadata Infrastructure resource to a different compartment.", ParameterSetName = Default)]
         public ChangeCompartmentDetails ChangeCompartmentDetails { get; set; }