+++ Marketplace/Cmdlets/Remove-OCIMarketplaceAcceptedAgreement.cs
@@ -21,7 +21,7 @@ namespace Oci.MarketplaceService.Cmdlets
         [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = @"The unique identifier for the accepted terms of use agreement.")]
         public string AcceptedAgreementId { get; set; }
 
-        [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, HelpMessage = @"Deprecated. The signature value is ignored.")]
+        [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, HelpMessage = @"Previously, the signature generated for the listing package terms of use agreement, but now deprecated and ignored.")]
         public string Signature { get; set; }
 
         [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, HelpMessage = @"Unique Oracle-assigned identifier for the request. If you need to contact Oracle about a particular request, please provide the request ID.")]